# api/swagger.yaml
swagger: '2.0'

info:
  version: "0.0.1"
  title: Magpie REST API
  description: The REST API used to interact with models hosted by a Magpie server

paths:
  /server/version:
    get:
      description: Get the version of the REST API
      responses:
        200:
          description: Successful response

  /server/status:
    get:
      description: |
       Information about the server. Includes the versions of the software and API, available memory, and processor count
      produces:
        - application/json
      responses:
        200:
          description: Successful response

  /model/{name}/info:
    get:
      description: |
        Gets information about a model stored in the server
      parameters:
        - name: name
          in: path
          description: Name of model
          required: true
          type: number
          format: int
      produces:
        - application/json
      responses:
        200:
          description: Successful response
        404:
          description: No model by that name found

  /model/{name}/model:
    get:
      description: |
        Get the model, serialized
      parameters:
        - name: name
          in: path
          description: Name of model
          required: true
          type: number
          format: int
      produces:
        - application/octet-stream
      responses:
        200:
          description: Successful response
        404:
          description: No model by that name found

  /model/{name}/dataset:
    get:
      description: |
        Get the dataset used to generate inputs for this model, serialized
      parameters:
        - name: name
          in: path
          description: Name of model
          required: true
          type: string
      produces:
        - application/octet-stream
      responses:
        200:
          description: Successful response
        404:
          description: No model by that name found

  /model/{name}/attributes:
    post:
      description: |
        Generate the attributes which serve as input into this model
      parameters:
        - name: name
          in: path
          description: Name of model
          required: true
          type: string
        - name: entries
          in: formData
          description: |
            Entries to be processed as JSON file. Only key in JSON file is
            "data," which contains the entries to be run written as strings
          type: string
          required: true
      produces:
        - application/json
      responses:
        200:
          description: Successful response
        405:
          description: Some problem in running the attribute generation

  /model/{name}/run:
    post:
      description: |
        Generate the attributes and run this model
      parameters:
        - name: name
          in: path
          description: Name of model
          required: true
          type: string
        - name: entries
          in: formData
          description: |
            Entries to be processed as JSON file. Only key in JSON file is
            "data," which contains the entries to be run written as strings
          type: string
          required: true
      produces:
        - application/json
      responses:
        200:
          description: Successful response
        405:
          description: Some problem in running the attribute generation